183 api tests rewrite (#186)
* created new test infrastructure and added a test for posting the person

* auto generate documentation

* added status code support

* updated api docs

* specify curl requests as shell code

* used printf instead of println

* updated flake.nix vendor hash

---------

+++ api/seekrd.go
@@ -52,7 +52,7 @@ func Birthday(db DataBase) DataBase {
 				}
 				log.Println(person.Birthday)
 				log.Println(age)
-				person.Age = age
+				person.Age = float64(age)
 			}
 		}
 		db[i] = person
